User commit

+++ src/main/java/entity/student.java
@@ -10,13 +10,13 @@ import jakarta.persistence.Id;
 public class Student {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private String name;
     private String email;
-    public int getId() {
+    public long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
     public String getName() {
@@ -31,7 +31,7 @@ public class Student {
     public void setEmail(String email) {
         this.email = email;
     }
-    public Student(int id, String name, String email, LocalDate dob, float cgpa) {
+    public Student(long id, String name, String email, LocalDate dob, float cgpa) {
         this.id = id;
         this.name = name;
         this.email = email;
